+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   IP address range, port range.
 - Publish check protocol compatibility with connected Programs.
 - Publish support log record data request/response through QUIC.
+- Add Packets Request GraphQL API.
 
 ### Changed